Add new page for current ItemList to create new Homepage (index.js)
