+++ components/Storages.js
@@ -1,0 +1,10 @@
+import Link from 'next/link';
+
+export default function Storages() {
+  return (
+    <section>
+      <h3>Fridge</h3>
+      <p>3 Items</p>
+    </section>
+  );
+}
