+++ lib/data.js
@@ -4,24 +4,28 @@ export const dummyitems = [
     name: 'Yogurt',
     date: '03/25/2034',
     quantity: '2',
+    storage: 'fridge',
   },
   {
     id: 2,
     name: 'Milk',
     date: '03/30/2034',
     quantity: '1',
+    storage: 'fridge',
   },
   {
     id: 3,
     name: 'Cheese',
     date: '03/20/2034',
     quantity: '1',
+    storage: 'fridge',
   },
   {
     id: 4,
     name: 'Eggs',
     date: '03/27/2034',
     quantity: '6',
+    storage: 'fridge',
   },
   {
     id: 5,
